services/ai/intent-parser/pyproject.toml
[project]
name = "intent-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "NitheshChakaravarthySeelan",email = "albertnitheesh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic[dotenv] (>=2.12.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "prometheus-client",
    "google-generativeai (>=0.7.0,<0.8.0)"
]

[tool.poetry.dev-dependencies]
black = "^24.4.2"
ruff = "^0.4.8"
pytest = "^8.2.2"

[tool.poetry]
packages = [{include = "intent_parser", from = "src"}]

[tool.poetry.scripts]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
